pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "microservicestest"
version = "0.1.0"
description = "Add your description here"
classifiers = ["Private :: Do Not Upload"]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.2.1",
    "duckdb-engine>=0.15.0",
    "fastapi>=0.115.8",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["api*"]
exclude = []
namespaces = false

[tool.ruff]
line-length = 99
extend-select = ["I"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
]
